Moved xlsx workbook and worksheet parsing to their own classes, mirroring the binary reader. Changed all parsers to inherit from the new internal interfaces IWorkbook and IWorksheet. Unified the common reader code into a new abstract ExcelDataReader<> base class inherited by the reader implementations. (#220)
